fix: types, formatting, unwanted committed files
