+++ src/node-tools/chunks_over_ws_server.ts
@@ -16,17 +16,12 @@ import { getWorldDemoEnv } from '../tests/configs/world_demo_setup.js'
 const SERVER_PORT = 3000
 const POOL_SIZE = 4
 
-const initWorkerpool = async () => {
+const initWsServer = async () => {
   const world_demo_env = getWorldDemoEnv()
   const chunks_node_worker_pool = new WorkerPool()
   await chunks_node_worker_pool.initPoolEnv(POOL_SIZE, world_demo_env)
   const chunks_scheduler = new ChunksPolling()
-  chunks_scheduler.chunksWorkerPool = chunks_node_worker_pool
-  return chunks_scheduler
-}
 
-const initWsServer = async () => {
-  const chunksScheduler = await initWorkerpool()
   const wsClients: Record<number, WebSocket> = {}
   const wss = new WebSocketServer({ port: SERVER_PORT })
 
@@ -35,12 +30,14 @@ const initWsServer = async () => {
     console.log('received client request:', request)
     const { viewPos, viewRange } = request
 
-    const scheduledTasks = chunksScheduler.pollChunks(
+    const chunks_tasks = chunks_scheduler.pollChunks(
       parseThreeStub(viewPos),
       viewRange,
     )
-    scheduledTasks?.forEach(scheduledTask =>
-      scheduledTask.then(chunkBlob => clientWs.send(chunkBlob)),
+    chunks_tasks?.forEach(chunks_task =>
+      chunks_task
+        .delegate(chunks_node_worker_pool)
+        .then(chunk_blob => clientWs.send(chunk_blob)),
     )
     // const clientTask = wsRequest.task
     // this.enqueueTasks(clientTask)
